issue-95: factor out apply effects
