+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Scott Gunther
   * Change searchNode.path to an array to support same action in path
   * Factor out checking of preconditions
+  * Factor out apply effects
 
 * Brian FItzpatrick
   * Added simulato walkthrough
