+++ lib/planner/planner-event-dispatch/register-planner-events.js
@@ -7,6 +7,7 @@ const searchNode = require('../search-node.js');
 const forwardStateSpaceSearch = require('../search-algorithms/forward-state-space-search-heuristic.js');
 const actionFocusedTechnique = require('../test-generatorion-techniques/action-focused-technique.js');
 const possibleActions = require('../possible-actions.js');
+const applyEffects = require('../apply-effects.js');
 
 module.exports = function(plannerEventDispatch) {
     plannerEventDispatch.on('planner.generateConfigured', testPlanner.generateTests);
@@ -62,6 +63,12 @@ module.exports = function(plannerEventDispatch) {
     forwardStateSpaceSearch.on('forwardStateSpaceSearch.getPossibleActions', function(...parameters) {
         process.nextTick(function() {
           possibleActions.emit('possibleActions.get', ...parameters);
+        });
+    });
+
+    forwardStateSpaceSearch.on('forwardStateSpaceSearch.applyEffects', function(...parameters) {
+        process.nextTick(function() {
+          applyEffects(...parameters);
         });
     });
