+++ lib/planner/search-algorithms/forward-state-space-search-heuristic.js
@@ -86,7 +86,7 @@ module.exports = forwardStateSpaceSearch = {
             clonedNode.path.push(action);
 
             yield forwardStateSpaceSearch
-                .emit('forwardStateSpaceSearch.applyActionToNode', clonedNode, next);
+                .emit('forwardStateSpaceSearch.applyEffects', clonedNode, next);
             let {applicableActions, allActions} = yield forwardStateSpaceSearch.
                 emit('forwardStateSpaceSearch.getPossibleActions', clonedNode, next);
             clonedNode.actions = applicableActions;
@@ -107,33 +107,6 @@ module.exports = forwardStateSpaceSearch = {
         }
 
         forwardStateSpaceSearch.emit('forwardStateSpaceSearch.findGoalActions');
-    },
-    _applyActionToNode(node, callback) {
-        let actionIdentifier = node.path[node.path.length -1];
-        let [name, actionName] = actionIdentifier.split('.');
-        let component = node.state.getComponentsAsMap().get(name);
-        let action = component.actions[actionName];
-
-        let testCaseAction = {
-            name: actionIdentifier,
-        };
-
-        if (Array.isArray(action.parameters)) {
-            let parameters = action.parameters.map(function(parameter) {
-                return parameter.generate.call(component, node.dataStore);
-            });
-            testCaseAction.options = {
-                parameters,
-            };
-            action.effects.call(component, ...testCaseAction.options.parameters, node.state, node.dataStore);
-        } else {
-            action.effects.call(component, node.state, node.dataStore);
-        }
-
-        node.testCase.push(testCaseAction);
-        node.lastAction = actionIdentifier;
-
-        callback();
     },
     * _testForGoal(node, callback) {
         let next = yield;
@@ -250,12 +223,6 @@ forwardStateSpaceSearch.on('forwardStateSpaceSearch.findGoalActions',
         });
     });
 
-forwardStateSpaceSearch.on('forwardStateSpaceSearch.applyActionToNode',
-    function(node, callback) {
-        process.nextTick(function() {
-            forwardStateSpaceSearch._applyActionToNode(node, callback);
-        });
-    });
 
 forwardStateSpaceSearch.on('forwardStateSpaceSearch.testForGoal',
     function(node, callback) {
